Figured some stuff out.
Figured out a better structure for things, think I am going to continue
to go with this until I figure out something better or see a better
example. This being adding switches for Set-DefaultApplications. I
wonder if there is a way that I can intergrate this with DSC? Also need
to think about a way that is non-interactive and more user friendly to
have granular control over settings.